auth-service/requirements.txt
fastapi
uvicorn
pydantic
python-jose
PyJWT
python-dotenv